Refactor ChannelOutboundBuffer to not use a cycular array to hold entries but use a linked list like structure
This is done by have DefaultChannelPromise implement the ChannelOutboundBuffer.Entry interface. This way the used ChannelPromise
for the write (the user is in almost all cases using DefaultChannelPromise) can be used as Entry and so no new objects needs to be created. This
helps to keep GC pressure to a minimum. For the situations where someone use a ChannelPromise which not implement ChannelOutboundBuffer.Entry
we use a RecyclableEntry to also reduce GC pressure. As part of this commit ChannelOutboundBuffer is not pooled anymore as this object
is pretty lightweight now.